Code/requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pillow==10.2.0
ply==3.11
psutil==5.9.8
pycryptodome==3.20.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0
tqdm==4.66.2
